aula5/requirements.txt
pymongo==3.8.0
python-dotenv==0.10.1
